+++ lua/plugins/conform.lua
@@ -21,7 +21,13 @@ return {
     -- Define your formatters
     formatters_by_ft = {
       lua = { "stylua" },
-      python = { "isort", "black" },
+      python = function(bufnr)
+        if require("conform").get_formatter_info("ruff", bufnr).available then
+          return { "ruff" }
+        else
+          return { "isort", "black" }
+        end
+      end,
       rust = { "rustfmt", lsp_format = "fallback" },
       -- javascript = { "prettierd", "prettier", stop_after_first = true },
       cs = { "csharpier" },